support hwtest and sdboot image generation
This patch adds a --deploy-all option for release image.
When we release a image, we need to build a sd bootable
image and hwtest image. The sdboot image will be located
in sdboot directory and hwtest images will be located
in hwtest directory. To generate a hw test image,
you have to locate manufacture-firmwares directory on top
of the directory.

Change-Id: Ibe8bea5f8fc55d9d1b00735037f86b48ddc487c2
Signed-off-by: Chanho Park <chanho61.park@samsung.com>