mksdfuse: use sector unit to correct sdboot image
This patch changes an unit of the fdisk to fix sdboot image since
applied the ota feature. Unlike the unit of the computer, the unit
of the disk is K:1000 and M:1000000. To fix broken offset issue of the
sdboot image, we have to change the unit of the mksdfuse script to
sector base.

Change-Id: Ib3306db686c0215908f13b0e09dae19cf5805206
Signed-off-by: Chanho Park <chanho61.park@samsung.com>