build_kernel: generate kernel module debug tarball
This patch is able to generate the kernel module debug tarball. It can
be used to debug kernel modules as well as vmlinux file. Before this
patch, there is no way to debug kernel modules. (Actually, the only way
to debug them is building the kernel on top of the same source revision)
This will generate two tarballs and they'll be used by debian package.

Change-Id: I8d0202a12dd8016db7eae6a766f7f3de4e77d8af
Signed-off-by: Chanho Park <chanho61.park@samsung.com>