ubuntu: do not archive .git directory
We're building a ubuntu package on top of the git source tree. However,
the dpkg-source which is called by sbuild did not exclude the .git
directory when generating a tarball. This patch fixes the tarball does
not contain the .git directory anymore. According to the dpkg-source
help, it has default patterns if there is no arguments after -I.
However, this patch only defines .git releated regular expressions
because I don't want to omit unwanted sources by the default patterns.

Change-Id: I39aeefd433c9f34c2284740a1a3f407fbe9e42a8
Signed-off-by: Chanho Park <chanho61.park@samsung.com>