build: prefer to use u-boot-dtb.bin
When we package U-BOOT binary in a chipset-specific format, prefer to
use u-boot-dtb.bin as a source image if it exists. In most cases of
ARTIK platforms where CONFIG_OF_SEPARATE is not set, u-boot-dtb.bin is
usually not available and u-boot.bin will be used instead.

This commit is a preparation step for enabling CONFIG_OF_SEPARATE.

Change-Id: Iba1416e188a60dae50bab2441c63c25b92686e0e
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>