mkvboot: do not put padding to DTB
mkimage sometimes returns an error after it failed to insert hashes or
signing signatures to an existing DTB. This actually happens if there
are not enough spaces between nodes in DTB. To prevent this situation,
we added some padding to reserve the spaces, passing additional options
to the device tree compiler through -D option.

Now that mkimage has been fixed to handle all such exceptional
conditions by expanding the DTB on the fly, we do not need to do so,
wasting spaces. So, let's remove it.

Change-Id: I1b1ac344ddb134d3a5449342b750eea14176e06d
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>