mkbootimg: fix VERIFIED_BOOT variable comparison
Fixes: d85cb1abd341 ("mkbootimg: populate fit image when verified boot
is enabled")

When we enabled the verified boot feature for kernel build, the
VERIFIED_BOOT variable can be defined by release.sh script. However,
the problem can be occurred when we're using the script alone without
the release.sh script. To fix this invalid comparison, we just compare
the variable "$VERIFIED_BOOT" == "true" at first. This patch also adds
--vboot option to enable the feature manually.

Change-Id: Ib3f8763d4892c53aeda304a00819481fb0d5f258
Signed-off-by: Chanho Park <chanho61.park@samsung.com>