release_rootfs: support to download next release rootfs
This patch can improve the download method of rootfs tarball. We can't
know which filename and md5 checksum will be generated before build. To
synchronize the source codes from in-house to github, we have to make
an additional patch to change the rootfs tarball name and md5 checksum.
Instead of changing it every release, we can estimate the release tag
from the OFFICIAL_RELEASE variable and we can get a file name from the
curl result of the URL(github URL + release tag) and md5 checksum as
well.

Change-Id: I073e31b6a30d258ada4be5c31adb50956a3fcee2
Signed-off-by: Chanho Park <chanho61.park@samsung.com>