build_uboot: prefer to use u-boot-dtb.bin
If u-boot-dtb.bin exists, use it as a source image. Otherwise, fallback
to u-boot.bin as before. We had already modified various build scripts
so that they perfer to use u-boot-dtb.bin, but there are still places
like this.

This change will not bring any functional differences until U-BOOT is
compiled with CONFIG_OF_SEPARATE=y.

Change-Id: I213c146b832824c8a9e2b2abc605d5231f297130
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>